Fizz Buzz test

+++ 0x04-more_functions_nested_loops/9-fizz_buzz.c
@@ -25,7 +25,7 @@ int main(void)
 		}
 		else
 		{
-			printf("%d", (int)a);
+			printf("%d", a);
 		}
 
 		if (a % 10 == 0)
